+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:places/data/favorites_repository/favorites_repository_memory.dart';
 import 'package:places/data/location_repository/location_repository_mock.dart';
-import 'package:places/data/sight_repository/sight_repository_memory.dart';
+import 'package:places/data/sight_repository/sight_repository_network.dart';
 import 'package:places/data/visited_repository/visited_repository_memory.dart';
 import 'package:places/domain/filter.dart';
 import 'package:places/domain/sight.dart';
@@ -13,7 +13,6 @@ import 'package:places/interactor/repository/visited_repository.dart';
 import 'package:places/interactor/settings_interactor/settings_interactor.dart';
 import 'package:places/interactor/sight_interactor.dart';
 import 'package:places/interactor/visiting_interactor.dart';
-import 'package:places/mocks.dart';
 import 'package:places/search_history_state.dart';
 import 'package:places/ui/res/app_strings.dart';
 import 'package:places/ui/res/themes.dart';
@@ -33,7 +32,8 @@ import 'package:places/ui/screen/visiting_screen.dart';
 final searchHistoryState = SearchHistoryState();
 
 // Временное место для интеракторов
-final SightRepository sightRepository = SightRepositoryMemory();
+// final SightRepository sightRepository = SightRepositoryMemory();
+final SightRepository sightRepository = SightRepositoryNetwork();
 final FavoritesRepository favoritesRepository = FavoritesRepositoryMemory();
 final VisitedRepository visitedRepository = VisitedRepositoryMemory();
 final LocationRepository locationRepository = LocationRepositoryMock();
@@ -59,7 +59,7 @@ final VisitedInteractor visitedInteractor = VisitedInteractor(
 final SettingsInteractor settingsInteractor = SettingsInteractor();
 
 Future<void> main() async {
-  await uploadMocks(sightRepository);
+  // await uploadMocks(sightRepository);
   runApp(App());
 }